% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{A function that takes a vector of numbers and returns Fizz if a number is
evenly divisible by 3 and/or Buzz if that number is evenly divisible by 5.}
\usage{
FizzBuzz(x)
}
\arguments{
\item{x}{A vector with real numbers}
}
\value{
The Fizz if the number in \code{x} is divisible by 3 and or Buzz if \code{x} is divisible by 5
}
\description{
A function that takes a vector of numbers and returns Fizz if a number is
evenly divisible by 3 and/or Buzz if that number is evenly divisible by 5.
}
\examples{
FizzBuzz(c(1,2,3))
FizzBuzz(5)
}
